+++ Platforms/Sniglet Widget/WidgetBundle.swift
@@ -8,9 +8,20 @@
 import SwiftUI
 import WidgetKit
 
-@main struct SnigletWidgets: WidgetBundle {
+@main
+struct SnigletWidgets: WidgetBundle {
+    @WidgetBundleBuilder
     var body: some Widget {
         RandomSniglet()
         DailySavedSniglet()
+        supportedWidgetsForFuture()
+    }
+
+    func supportedWidgetsForFuture() -> some Widget {
+        if #available(iOS 16.0, *) {
+            return WidgetBundleBuilder.buildBlock(RandomSnigletAccessory())
+        } else {
+            return WidgetBundleBuilder.buildBlock()
+        }
     }
 }