+++ Shared/Utilities/AVSpeechSynthesizer.swift
@@ -1,0 +1,12 @@
+//
+//  AVSpeechSynthesizer.swift
+//  Give Me a Sniglet
+//
+//  Created by Marquis Kurt on 9/14/22.
+//
+
+import AVFoundation
+
+extension AVSpeechSynthesizer {
+    public static let shared = AVSpeechSynthesizer()
+}